requirements.txt
openai
python-docx
beautifulsoup4
requests
faiss-cpu
tqdm
email-validator
python-dotenv
imap-tools
flask
jinja2
pypdf
werkzeug
cryptography